+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - controller: allow overriding some build-time environment variables ([#137](https://git.opentalk.dev/opentalk/k3k-controller/-/issues/137))
 - chat: add `last_seen_timestamp` fields [#242](https://git.opentalk.dev/opentalk/k3k-controller/-/issues/242)
+- legalvote: add option to automatically create a PDF asset when a vote has ended ([#259](https://git.opentalk.dev/opentalk/k3k-controller/-/issues/259))
 
 ### Fixed